requirements.txt
djangorestframework
djoser
drf_writable_nested
